requirements.txt
flask
torch
numpy
soundfile
noisereduce
pydub
openai-whisper
transformers
TTS